move middleware to separate files
